// __tests__/sdk/agentsmith.types.ts
/*
|--------------------------------------------|
         ⚡️ Generated by Agentsmith ⚡️
      ⚠️ Do not edit this file directly ⚠️
|--------------------------------------------|
*/
export type Agency = {
  prompts: {
    'support-chat': {
      uuid: '33c3c066-be82-4f15-9415-390bae2def53';
      name: 'Support Chat';
      slug: 'support-chat';
      versions: {
        latest: {
          uuid: '5d714a31-3e65-4b19-b6c6-12aaa274f417';
          version: '0.1.0';
          config: any;
          content: string;
          variables: { userMessage: string };
        };
        '0.1.0': {
          uuid: '5d714a31-3e65-4b19-b6c6-12aaa274f417';
          version: '0.1.0';
          config: any;
          content: string;
          variables: { userMessage: string };
        };
        '0.0.1': {
          uuid: '55b88216-c8f1-4b1c-9bac-2b92a54f2ff4';
          version: '0.0.1';
          config: any;
          content: string;
          variables: { userMessage: string };
        };
      };
    };
    'hello-world': {
      uuid: '5b3a72a5-4af6-4670-a18f-0ce43035dce4';
      name: 'Hello World';
      slug: 'hello-world';
      versions: {
        latest: {
          uuid: '4d2bdcb8-d8f9-4a32-ba1e-ffd8d9acf06a';
          version: '0.0.2';
          config: any;
          content: string;
          variables: { name: string };
        };
        '0.0.2': {
          uuid: '4d2bdcb8-d8f9-4a32-ba1e-ffd8d9acf06a';
          version: '0.0.2';
          config: any;
          content: string;
          variables: { name: string };
        };
        '0.0.1': {
          uuid: 'aac26f05-3d14-45a6-bded-0c8de072f796';
          version: '0.0.1';
          config: any;
          content: string;
        };
      };
    };
  };
  globals: {
    gitHubUrl: 'https://github.com/chad-syntax/agentsmith';
    supportUrl: 'https://agentsmith.app/support';
    companyName: 'Agentsmith';
    supportEmail: 'support@agentsmith.app';
  };
};